SportsVision/requirements.txt
# SportsVision Requirements
# Install with: pip install -r requirements.txt

# Core ML
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
pandas>=2.0.0

# PyTorch (install via conda for CUDA support)
# conda install pytorch torchvision torchaudio pytorch-cuda=11.8 -c pytorch -c nvidia
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Model visualization (optional)
torchinfo>=1.8.0
torchview>=0.2.6

# Training utilities
tqdm>=4.65.0
wandb>=0.15.0

# Configuration
PyYAML>=6.0

# Visualization
matplotlib>=3.7.0

# Streamlit demo app
streamlit>=1.28.0

# JSON handling (standard library, listed for clarity)
# json - built-in

# Optional: for development
# ipython>=8.0.0
# jupyter>=1.0.0
